Fix circular dependencies
- Moved the dropdown-focus-handler.directive inside the dropdown.component since they used one another which
caused a circular dependency
- Moved 'lastAvailabilityValue' constant from action-menu.component to constants.ts because it was used by
the dropdown.component which was used by the action-menu.component
- Moved 'NonEmptyArray' type from shared-modal.component into constants.ts since it was used
by the shared-modal-tab.component which was used by the shared modal.component
