+++ projects/components/src/dropdown/dropdown.module.ts
@@ -9,7 +9,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { ClarityModule } from '@clr/angular';
 import { I18nModule } from '@vcd/i18n';
 import { ShowClippedTextDirectiveModule } from '../lib/directives/show-clipped-text.directive.module';
-import { DropdownFocusHandlerDirective } from './dropdown-focus-handler.directive';
+import { DropdownFocusHandlerDirective } from './dropdown.component';
 import { DropdownComponent } from './dropdown.component';
 import { DynamicDropdownPositionDirective } from './dynamic-dropdown-position.directive';
